test(infra): split agent run task ownership coverage
