+++ src/infra/agent-events.test.ts
@@ -17,11 +17,9 @@ import {
   withAgentRunLifecycleGeneration,
 } from "./agent-events.js";
 import {
-  bindAgentRunTaskRunId,
   claimAgentRunContext,
   clearAgentRunContext,
   getAgentRunContext,
-  getAgentRunTaskRunId,
   listAgentRunsForSession,
   readAgentRunIndexVersion,
   registerAgentRunContext,
@@ -78,39 +76,6 @@ describe("agent-events sequencing", () => {
     expect(getAgentRunContext("run-1")?.sessionKey).toBe("main");
     clearAgentRunContext("run-1");
     expect(getAgentRunContext("run-1")).toBeUndefined();
-  });
-
-  test("binds detached task ids to exact active claims", () => {
-    const lifecycleGeneration = getAgentEventLifecycleGeneration();
-    const firstClaim = claimAgentRunContext(
-      "shared-task-run",
-      { sessionKey: "agent:main:cron-task-session", lifecycleGeneration },
-      { trackOwner: true, ownsContext: true },
-    );
-    const secondClaim = claimAgentRunContext(
-      "shared-task-run",
-      { lifecycleGeneration },
-      { trackOwner: true, ownsContext: true },
-    );
-    expect(firstClaim).toBeTruthy();
-    expect(secondClaim).toBeTruthy();
-    if (!firstClaim || !secondClaim) {
-      throw new Error("expected tracked agent run claims");
-    }
-
-    expect(bindAgentRunTaskRunId("shared-task-run", "missing-claim", "task-a")).toBe(false);
-    expect(bindAgentRunTaskRunId("shared-task-run", firstClaim, "  task-a  ")).toBe(true);
-    expect(bindAgentRunTaskRunId("shared-task-run", secondClaim, "   ")).toBe(false);
-    expect(getAgentRunTaskRunId("shared-task-run")).toBe("task-a");
-
-    expect(bindAgentRunTaskRunId("shared-task-run", secondClaim, "task-b")).toBe(true);
-    expect(getAgentRunTaskRunId("shared-task-run")).toBeUndefined();
-
-    releaseAgentRunContext("shared-task-run", secondClaim);
-    expect(getAgentRunTaskRunId("shared-task-run")).toBe("task-a");
-    releaseAgentRunContext("shared-task-run", firstClaim);
-    expect(getAgentRunTaskRunId("shared-task-run")).toBeUndefined();
-    expect(getAgentRunContext("shared-task-run")).toBeUndefined();
   });
 
   test("versions active-run projection ownership transitions", () => {
